chore: move google drive client under the googledrive pkg
