+++ pkg/component/data/googledrive/v0/main_test.go
@@ -39,6 +39,8 @@ var (
 	sharedFolderLink = fmt.Sprintf("https://drive.google.com/drive/folders/%s?usp=drive_link", fakeID)
 )
 
+// We will unify the unit tests for the component development.
+
 func Test_Execute_ReadFile(t *testing.T) {
 
 	c := qt.New(t)
